include super_selectable in all controllers
